【新增】首页文章获取（Paging3） 【修复】Json序列化时出现的严格控制字段缺省值等问题

+++ app/src/main/java/com/zjl/finalarchitecture/module/home/model/PageVO.kt
@@ -13,7 +13,7 @@ data class PageVO<out T>(
      * 当前页码
      */
     @SerialName("curPage")
-    val currentPage: Long,
+    val currentPage: Int,
     /**
      * 数据列表
      */
